update Comment
eliminarComentario
handleEditComentario
cambiar el botón de agregar comentario a handleedit
cambiar el data 3 variables al final
updatecomentario

+++ src/services/CharlasService.js
@@ -364,13 +364,16 @@ export default class CharlasService {
         return new Promise((resolve, reject) => {
             const endpoint = `api/comentarios/${idComentario}`;
             const token = Cookies.get('bearer_token');
-            
+
+            if (!token) {
+                reject("Token de autenticación no encontrado.");
+                return;
+            }
             axios.delete(
                 Global.urlBase + endpoint,
                 {
                     headers: {
-                        Authorization: `Bearer ${token}`, 
-                        'Content-Type': 'application/json'
+                        Authorization: token,
                     }
                 }
             )
@@ -378,8 +381,8 @@ export default class CharlasService {
                 resolve(response.data);
             })
             .catch(error => {
-                console.error("Error al eliminar el comentario: ", error.response ? error.response.data : error);
-                reject(error);
+                console.error("Error al eliminar el comentario:", error.response ? error.response.data : error);
+                reject(error.response ? error.response.data : error);
             });
         });
     }
@@ -444,6 +447,19 @@ export default class CharlasService {
             });
         });
     }
+
+    updateComentario(comentario){
+		const request = "api/comentarios";
+		const url = Global.urlBase + request;
+        const token = Cookies.get('bearer_token');
+        console.log(token);
+
+		return axios.put(url, comentario, {
+			headers: {
+				'Authorization': token
+			}
+		})
+	}
 
     editarCharla(charlaActualizada) {
         return new Promise((resolve, reject) => {
